making CallableLoader generic
